Glasp: An Effective Tool For Writers
Glasp is a tool that helps people highlight keynotes, add comments, share, and have a quote shot of their keynotes to share with other like-minds. 

Features
Social note-taking
Access to a collection of the highlighted text
Sharing of keynotes with like-minds.

Signing up
Signing up on Glasp is free and open to all users.

Conclusion
Glasp has been tested and trusted by people. It is a solution provider to the traditional means of note-taking without having to go back and forth between screen