Glasp: An Effective Tool For Writers
Glasp is a helpful and effective tool for people, especially writers, and researchers. It helps them to highlight keynotes, add comments, and share their ideas with like-minds.

Features
Social notetaking
Access to the collection of the highlighted text
Sharing of ideas and insight with like-minds

Signing up
Signing up for Glasp is free and it is open to all users.

Conclusion
Glasp has been an effective tool for all its users. It allows them to connect and share insights on specific topics. It overwrites the traditional way of taking notes without having to go back and forth between screens.